refactor: add type dir with types and interfaces
