+++ src/components/types/interfaces.ts
@@ -1,0 +1,9 @@
+interface IController {
+  fetchRecent(): void;
+  fetchSearchCall(search: string): void;
+  fetchById(id: number): void;
+  fetchEpisodesById(id: number): void;
+  test(): void;
+}
+
+export {IController};
